Move parsing of color names to `named` module
